+++ tests/testthat/test-create.R
@@ -20,6 +20,28 @@ test_that("dir_create works with new and existing files", {
   unlink(x1, recursive = TRUE)
 })
 
+test_that("file_create works with multiple path arguments", {
+  x1 <- file_create(tempdir(), "foo")
+
+  expect_equal(x1, path(tempdir(), "foo"))
+  expect_true(file_exists(x1))
+
+  unlink(x1)
+})
+
+test_that("dir_create works with multiple path arguments", {
+  x1 <- dir_create(tempdir(), "foo", "bar")
+  x2 <- dir_create(tempdir(), "foo", "baz", recurse = FALSE)
+
+  expect_equal(x1, path(tempdir(), "foo", "bar"))
+  expect_equal(x2, path(tempdir(), "foo", "baz"))
+  expect_true(dir_exists(x1))
+  expect_true(dir_exists(x2))
+
+  unlink(x1)
+  unlink(x2)
+})
+
 test_that("dir_create sets the mode properly", {
   skip_on_cran()
   skip_on_os("windows")